+++ packages/spacecat-shared-data-access/test/fixtures/index.fixtures.js
@@ -23,6 +23,7 @@ import keyEvents from './key-events.fixture.js';
 import opportunities from './opportunities.fixture.js';
 import organizations from './organizations.fixture.js';
 import siteCandidates from './site-candidates.fixture.js';
+import siteTopForms from './site-top-forms.fixture.js';
 import siteTopPages from './site-top-pages.fixture.js';
 import sites from './sites.fixture.js';
 import suggestions from './suggestions.fixture.js';
@@ -43,6 +44,7 @@ export default {
   opportunities,
   organizations,
   siteCandidates,
+  siteTopForms,
   siteTopPages,
   sites,
   suggestions,