+++ packages/spacecat-shared-utils/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@adobe/spacecat-shared-utils-v1.44.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.44.0...@adobe/spacecat-shared-utils-v1.44.1) (2025-07-28)
+
+
+### Bug Fixes
+
+* expose calendar week helper function ([#881](https://github.com/adobe/spacecat-shared/issues/881)) ([5e178a8](https://github.com/adobe/spacecat-shared/commit/5e178a8eeaf426612f593f016f765b76cd565c03))
+
+# [@adobe/spacecat-shared-utils-v1.44.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.43.1...@adobe/spacecat-shared-utils-v1.44.0) (2025-07-28)
+
+
+### Features
+
+* add function for extracting list of calendar weeks ([#880](https://github.com/adobe/spacecat-shared/issues/880)) ([f69a808](https://github.com/adobe/spacecat-shared/commit/f69a808b96ee5e977e5cf5936e3effa378b51581))
+
+# [@adobe/spacecat-shared-utils-v1.43.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.43.0...@adobe/spacecat-shared-utils-v1.43.1) (2025-07-27)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#878](https://github.com/adobe/spacecat-shared/issues/878)) ([b049828](https://github.com/adobe/spacecat-shared/commit/b04982839c0ff5e4de4ab0e37508c5eb5272a679))
+
 # [@adobe/spacecat-shared-utils-v1.43.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-utils-v1.42.4...@adobe/spacecat-shared-utils-v1.43.0) (2025-07-23)